improve variable types

+++ src/controllers/ElevationController.h
@@ -6,15 +6,15 @@
 
 struct ElevationControllerConfig
 {
-  int motorEnPin;
-  int motorPwmUpPin;
-  int motorPwmDownPin;
-  int motorPwmSpeed;
+  uint8_t motorEnPin;
+  uint8_t motorPwmUpPin;
+  uint8_t motorPwmDownPin;
+  uint8_t motorPwmSpeed;
   float maxAzimuth;
   float minAzimuth;
   float maxElevation;
   float minElevation;
-  unsigned long timeThreshold;
+  uint32_t timeThreshold;
   float actuatorSpeed;
   float actuatorLength;
 };
@@ -38,10 +38,10 @@ public:
 
 private:
   // Pin Definitions
-  int motorPinEn;      // Motor enable pin
-  int motorPinPwmUp;   // Motor PWM pin for actuator extension (up)
-  int motorPinPwmDown; // Motor PWM pin for actuator retraction (down)
-  int motorPwmSpeed;   // Motor PWM speed
+  uint8_t motorPinEn;      // Motor enable pin
+  uint8_t motorPinPwmUp;   // Motor PWM pin for actuator extension (up)
+  uint8_t motorPinPwmDown; // Motor PWM pin for actuator retraction (down)
+  uint8_t motorPwmSpeed;   // Motor PWM speed
 
   // Motor Controller
   BTS7960 motorController;
@@ -50,17 +50,17 @@ private:
   float currentElevation; // Current elevation position of the solar panel
 
   // Actuator parameters
-  float actuatorSpeed;                  // Speed of the actuator in mm/s
-  float actuatorLength;                 // Length of the actuator in mm
-  unsigned long actuatorFullTravelTime; // Time in milliseconds to move the actuator from fully retracted to fully extended
-  float degreesPerMs;                   // Degrees the actuator moves per millisecond
+  float actuatorSpeed;             // Speed of the actuator in mm/s
+  float actuatorLength;            // Length of the actuator in mm
+  uint32_t actuatorFullTravelTime; // Time in milliseconds to move the actuator from fully retracted to fully extended
+  float degreesPerMs;              // Degrees the actuator moves per millisecond
 
   // Constants
   const float azimuthDegMax;
   const float azimuthDegMin;
   const float elevationDegMax;
   const float elevationDegMin;
-  const unsigned long elevationTimeThreshold;
+  const uint32_t elevationTimeThreshold;
 };
 
 extern ElevationController elevationController;
